linked buttons on homepage container to appropriate pages with the exception of fullschedule button-that needs to go to the full schedule page that is not yet created. right now it's going to appointments card
